% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartwig.R
\docType{data}
\name{Hartwig}
\alias{Hartwig}
\title{Hartwig Medical Foundation (Hartwig)}
\format{
An object of class \code{"phyloseq"}.
}
\source{
https://www.hartwigmedicalfoundation.nl/en/
}
\usage{
data('Hartwig')
}
\description{
Mapping-based analysis of microbial profiling from metastatic tumors
}
\details{
WGS unmapped reads with Kraken2/PathSeq profiling from metastatic tumors
}
\references{
Battaglia  et al. (2024) Cell
(\href{}{PubMed})
}
\keyword{datasets}
