% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMPv35.R
\docType{data}
\name{HMPv35}
\alias{HMPv35}
\title{HMP 16S rRNA data from V3-V5}
\format{An object of class \code{"phyloseq"}.}
\usage{
data('HMPv35')
}
\description{
This dataset contains 16S rRNA amplicon data from the HMP Project.
}
\references{
The NIH HMP Working Group et al. (2009) Genome Res 19(12): 2317–2323.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2792171}{PubMed})
}
\keyword{datasets}

