% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovingPictures.R
\docType{data}
\name{MovingPictures}
\alias{MovingPictures}
\title{Moving pictures of the human microbiome (Qiita-550)}
\format{An object of class \code{"phyloseq"}.}
\source{
https://qiita.ucsd.edu/study/description/550
}
\usage{
data('MovingPictures')
}
\description{
Understanding the normal temporal variation in the human microbiome is critical to developing treatments for putative microbiome-related afflictions such as obesity, Crohn's disease, inflammatory bowel disease and malnutrition. Sequencing and computational technologies, however, have been a limiting factor in performing dense time series analysis of the human microbiome. Here, we present the largest human microbiota time series analysis to date, covering two individuals at four body sites over 396 timepoints.
}
\details{
Closed-reference OTU-picking with SortMeRNA (97 percent identity)
}
\references{
Caporaso et al. (2009) Genome Biol. 2011;12(5):R50.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/21624126}{PubMed})
}
\keyword{datasets}

