% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISK_CCFA.R
\docType{data}
\name{RISK_CCFA}
\alias{RISK_CCFA}
\title{The Treatment-Naive Microbiome in New-Onset Crohn s Disease (Qiita-1939)}
\format{An object of class \code{"phyloseq"}.}
\usage{
data('RISK_CCFA')
}
\description{
Understanding the normal temporal variation in the human microbiome is critical to developing treatments for putative microbiome-related afflictions such as obesity, Crohn's disease, inflammatory bowel disease and malnutrition. Sequencing and computational technologies, however, have been a limiting factor in performing dense time series analysis of the human microbiome. Here, we present the largest human microbiota time series analysis to date, covering two individuals at four body sites over 396 timepoints.
}
\details{
Closed-reference OTU-picking with SortMeRNA (97\% identity)
}
\references{
Gevers et al. (2014) Cell Host Microbe. 2014 Mar 12;15(3):382-92.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/21624126}{PubMed})
}
\keyword{datasets}

