% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\docType{data}
\name{TCGA}
\alias{TCGA}
\title{The Cancer Genome Atlas (TCGA)}
\format{
An object of class \code{"phyloseq"}.
}
\source{
ftp://ftp.microbio.me/pub/cancer_microbiome_analysis
}
\usage{
data('TCGA')
}
\description{
The Cancer Genome Atlas (TCGA) of 33 types of cancer from treatment-naive patients (a total of 18,116 samples) for microbial reads
}
\details{
WGS unampped reads with Kraken2 profiling from primary tumors
}
\references{
Poore  et al. (2020) Nature Mar;579(7800):567-574
(\href{https://pubmed.ncbi.nlm.nih.gov/32214244/}{PubMed})
}
\keyword{datasets}
