% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMPv35.R
\docType{data}
\name{HMPv35}
\alias{HMPv35}
\title{HMP 16S rRNA data from V3-V5}
\format{An object of class \code{"phyloseq"}.}
\usage{
data('HMPv35')
}
\description{
The Human Microbiome Project (HMP), funded as an initiative of the NIH Roadmap for Biomedical Research (http://nihroadmap.nih.gov), is a multi-component community resource. The goals of the HMP are: (1) to take advantage of new, high-throughput technologies to characterize the human microbiome more fully by studying samples from multiple body sites from each of at least 250 “normal” volunteers; (2) to determine whether there are associations between changes in the microbiome and health/disease by studying several different medical conditions; and (3) to provide both a standardized data resource and new technological approaches to enable such studies to be undertaken broadly in the scientific community. The ethical, legal, and social implications of such research are being systematically studied as well. The ultimate objective of the HMP is to demonstrate that there are opportunities to improve human health through monitoring or manipulation of the human microbiome. The history and implementation of this new program are described here.
}
\details{
Closed-reference OTU-picking with UCLUST (97\% identity)
}
\references{
The NIH HMP Working Group et al. (2009) Genome Res 19(12): 2317–2323.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2792171}{PubMed})
}
\keyword{datasets}

